fix (core module): bug fix "circular dependency"; merged authenticationService and usersService
