--- projects/knora/core/src/lib/services/authentication/authentication.service.spec.ts
@@ -1,209 +1,0 @@
-import {async, inject, TestBed} from '@angular/core/testing';
-import {ApiService} from './api.service';
-import {ProjectsService} from './projects.service';
-import {HttpClient, HttpClientModule} from '@angular/common/http';
-import {ApiServiceError, Project, ProjectResponse, ProjectsResponse} from '../declarations';
-import {
-    anythingProject,
-    anythingProjectResponseJson,
-    imagesProject,
-    imagesProjectResponseJson, incunabulaProject,
-    incunabulaProjectResponseJson,
-    projectsResponseJson,
-    projectsTestData
-} from '../test-data/shared-test-data';
-import {JsonConvert, OperationMode, ValueCheckingMode} from 'json2typescript';
-
-
-describe('ProjectsService', () => {
-    beforeEach(() => {
-        TestBed.configureTestingModule({
-            providers: [
-                ApiService,
-                ProjectsService,
-                HttpClient
-            ]
-        });
-    });
-
-    it('should be created', async(inject(
-        [ProjectsService], (service) => {
-            expect(service).toBeDefined();
-        }))
-    );
-
-
-    it('should parse projects-response', () => {
-
-        const jsonConvert: JsonConvert = new JsonConvert(OperationMode.ENABLE, ValueCheckingMode.ALLOW_NULL);
-
-        const result: ProjectsResponse = jsonConvert.deserializeObject(projectsResponseJson, ProjectsResponse);
-
-        expect(result).toBeTruthy();
-    });
-
-    it('should parse project-response (images)', () => {
-
-        const jsonConvert: JsonConvert = new JsonConvert(OperationMode.ENABLE, ValueCheckingMode.ALLOW_NULL);
-
-        const result: ProjectResponse = jsonConvert.deserializeObject(imagesProjectResponseJson, ProjectResponse);
-
-        expect(result).toBeTruthy();
-    });
-
-    it('should parse project-response (incunabula)', () => {
-
-        const jsonConvert: JsonConvert = new JsonConvert(OperationMode.ENABLE, ValueCheckingMode.ALLOW_NULL);
-
-        const result: ProjectResponse = jsonConvert.deserializeObject(incunabulaProjectResponseJson, ProjectResponse);
-
-        expect(result).toBeTruthy();
-    });
-
-    it('should parse project-response (anything)', () => {
-
-        const jsonConvert: JsonConvert = new JsonConvert(OperationMode.ENABLE, ValueCheckingMode.ALLOW_NULL);
-
-        const result: ProjectResponse = jsonConvert.deserializeObject(anythingProjectResponseJson, ProjectResponse);
-
-        expect(result).toBeTruthy();
-    });
-
-
-    /*
-    if (environment.type === 'integration') {
-
-        it('should load test data [it]', async(inject(
-            [StoreService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.resetTriplestoreContent([])
-                    .subscribe(
-                        (result: string) => {
-                            expect(result).toBe('success');
-                        });
-
-            })), 300000);
-
-
-        it('#getAllProjects should return all projects [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getAllProjects()
-                    .subscribe(
-                        (projects: Project[]) => {
-                            // console.log('projects: ' + JSON.stringify(projects));
-                            expect(projects.length).toBe(8);
-                            expect(projects).toEqual(projectsTestData);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-
-        it('#getProjectByIri should return project (images) [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getProjectByIri('http://rdfh.ch/projects/00FF')
-                    .subscribe(
-                        (project: Project) => {
-                            // console.log('project: ' + JSON.stringify(project));
-                            expect(project).toEqual(imagesProject);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-        it('#getProjectByIri should return project (incunabula) [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getProjectByIri('http://rdfh.ch/projects/77275339')
-                    .subscribe(
-                        (project: Project) => {
-                            // console.log('project: ' + JSON.stringify(project));
-                            expect(project).toEqual(incunabulaProject);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-        it('#getProjectByIri should return project (anything) [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getProjectByIri('http://rdfh.ch/projects/anything')
-                    .subscribe(
-                        (project: Project) => {
-                            // console.log('project: ' + JSON.stringify(project));
-                            expect(project).toEqual(anythingProject);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-        it('#getProjectByShortname should return project (images) [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getProjectByShortname('images')
-                    .subscribe(
-                        (project: Project) => {
-                            // console.log('users: ' + JSON.stringify(users));
-                            expect(project).toEqual(imagesProject);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-
-        it('#getProjectByShortcode should return project (images) [it]', async(inject(
-            [ProjectsService], (service) => {
-
-                expect(service).toBeDefined();
-
-                service.getProjectByShortcode('00FF')
-                    .subscribe(
-                        (project: Project) => {
-                            // console.log('users: ' + JSON.stringify(users));
-                            expect(project).toEqual(imagesProject);
-                        },
-                        (error: ApiServiceError) => {
-                            fail(error);
-                        }
-                    );
-
-            })));
-
-
-
-    } else {
-        xit('integration tests skipped. run  "ng test --env=it".');
-    }
-    */
-
-
-});
