+++ projects/knora/core/src/lib/services/index.ts
@@ -8,4 +8,3 @@ export * from './api.service';
  */
 export * from './projects/projects.service';
 export * from './users/users.service';
-export * from './authentication/authentication.service';
