+++ projects/knora/core/src/public_api.ts
@@ -2,9 +2,9 @@
  * Public API Surface of core
  */
 
-export * from './lib/declarations/';
 export * from './lib/core.module';
+export * from './lib/declarations/';
+export * from './lib/services';
 
-export * from './lib/services/projects/projects.service';
-export * from './lib/services/users/users.service';
-export * from './lib/services/authentication/authentication.service';
+// export * from './lib/services/projects/projects.service';
+// export * from './lib/services/users/users.service';
